Rewrite get_next_match_region to be faster and easier.
This change rewrites get_next_match_region to use a "candidate"
variable. This allows us to make a linear scan of the input without
jumping forward and scanning the same byte twice. This change also
improves the match selection heuristic to look at the distance when
considering two matches that are close to one another. This gives a
small 1% performance improvement and a 1% size improvement.